refactor(spindrift): one read hook, and the screens beside their views
`usePoll` answered the cadence question and left every screen to hold the
read: a three-armed `useState`, a `Promise.all`, an `ok` check per result,
an aborted guard, a catch that made a sentence, and a token to re-run it.
`useRead` answers all of it once — the first refusal wins, a refusal
replaces the screen and a throw keeps what is readable, and `merge` is
there for the two ledgers and the workspace, whose fresh read is not the
whole truth. `connections.tsx`'s own `useConnection` was the same hook,
shallower; it is gone.

The fourteen screens move out of `app.tsx` beside the views they render,
taking their skeletons with them. What is left is the shell, the two
gating states, the route table and the chrome — 3042 lines to 477.

The route table's ordered `startsWith` chain is unchanged: a
longest-prefix table cannot express its two exact matches, its id-or-list
splits, or the bare-path-is-an-App catch-all without growing.
