+++ apps/spindrift/test/commands/component-added-beside-a-sibling.test.ts
@@ -13,7 +13,7 @@
  *   (`src/commands/apps/deploy.ts:529-534`) and a form that placed as well
  *   would be a second answer to which Target this Component lives on;
  * - its first Deploy therefore has to be told a Target, which is the one
- *   conditional the screen adds (`src/web/app.tsx`'s `handleDeploy`), and the
+ *   conditional the screen adds (`src/web/views/apps/workspace.tsx`'s `handleDeploy`), and the
  *   Build that press starts is the new Component's own.
  */
 import { describe, expect, test } from 'bun:test';
