+++ apps/spindrift/test/web/workspace-refresh.test.ts
@@ -14,7 +14,7 @@
  */
 import { describe, expect, test } from 'bun:test';
 import type { WorkspaceView } from '../../src/commands/views.ts';
-import { refreshedWorkspace } from '../../src/web/app.tsx';
+import { refreshedWorkspace } from '../../src/web/views/apps/workspace.tsx';
 import { WORKSPACE_SCENARIOS } from '../fixtures/scenarios.ts';
 
 const SERVICE: WorkspaceView = WORKSPACE_SCENARIOS.service;
